refactor range iterator to core
